refactor: move useDeletePostDialog custom hook from model to lib folder
